extracted different extension groups to separate files
